feat(ui): save a screenshot of the interface with Ctrl+S
A complete SVG renderer for a drawn terminal buffer already existed inside
`tui_tester`, where it generates the images in the README. The application
itself had no way to capture what it was showing, which is the first thing
asked for in a bug report about a TUI.

Moved to `ui::screenshot` and shared by both. Ctrl+S writes the frame that
was just drawn to `<data>/screenshots/<timestamp>.svg` and reports the path
in the status line. SVG rather than a bitmap keeps the text selectable and
the file small, and needs no image encoder.

Named by timestamp so repeated presses accumulate: the point is usually to
capture a sequence, or something that just happened and may not repeat.

Handled at the top level rather than in a tab, since the buffer belongs to
the terminal and not to any one screen. Ctrl+Ы is bound alongside Ctrl+S
for Russian layouts, as with the other shortcuts.

Verified that `tui_tester` still generates all 14 README screenshots from
the shared code, byte-identical to the ones already committed.
