+++ tui/src/bin/tui_tester.rs
@@ -5,141 +5,12 @@ use ac_core::engineer::{Recommendation, Severity};
 use ac_core::overlay::OverlayMode;
 use ac_core::session_info::SessionInfo;
 use ac_tui::ui::UIRenderer;
+use ac_tui::ui::screenshot::buffer_to_svg;
 use ac_tui::{AppStage, AppState, AppTab, SafeLock};
 use ratatui::Terminal;
 use ratatui::backend::TestBackend;
 use std::fs;
 use std::path::Path;
-
-fn escape_xml(s: &str) -> String {
-    s.replace('&', "&amp;")
-        .replace('<', "&lt;")
-        .replace('>', "&gt;")
-        .replace('"', "&quot;")
-        .replace('\'', "&apos;")
-}
-
-fn buffer_to_svg(
-    buffer: &ratatui::buffer::Buffer,
-    width: u16,
-    height: u16,
-    output_path: &Path,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let char_w = 10u32;
-    let char_h = 20u32;
-    let header_h = 32u32;
-    let font_size = 14u32;
-    let content_w = (width as u32) * char_w;
-    let content_h = (height as u32) * char_h;
-    let total_w = content_w + 16;
-    let total_h = content_h + header_h + 16;
-
-    let mut svg = String::with_capacity(64 * 1024);
-    svg.push_str(&format!(
-        r#"<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 {} {}" width="{}" height="{}" style="background-color: #0b0f19; font-family: 'Fira Code', 'JetBrains Mono', 'Cascadia Code', 'Consolas', monospace; font-size: {}px;">"#,
-        total_w, total_h, total_w, total_h, font_size
-    ));
-    svg.push('\n');
-
-    // Outer Card Frame with Shadow & Border
-    svg.push_str(&format!(
-        r##"  <rect x="4" y="4" width="{}" height="{}" rx="10" fill="#0d1117" stroke="#30363d" stroke-width="1.5"/>"##,
-        total_w - 8, total_h - 8
-    ));
-    svg.push('\n');
-
-    // Window Control Buttons (Red, Yellow, Green)
-    svg.push_str(r##"  <circle cx="20" cy="20" r="5.5" fill="#ff5f56"/>"##);
-    svg.push('\n');
-    svg.push_str(r##"  <circle cx="36" cy="20" r="5.5" fill="#ffbd2e"/>"##);
-    svg.push('\n');
-    svg.push_str(r##"  <circle cx="52" cy="20" r="5.5" fill="#27c93f"/>"##);
-    svg.push('\n');
-
-    // Window Title Text
-    svg.push_str(&format!(
-        r##"  <text x="{}" y="24" fill="#8b949e" font-size="12px" font-weight="bold" text-anchor="middle">AC Pro Engineer v{} — High-Performance Sim Telemetry Suite</text>"##,
-        total_w / 2,
-        // Was hardcoded to v0.2.3, so every screenshot in the README claimed
-        // a version two releases old.
-        ac_core::updater::CURRENT_VERSION
-    ));
-    svg.push('\n');
-
-    // Divider Line
-    svg.push_str(&format!(
-        r##"  <line x1="4" y1="{}" x2="{}" y2="{}" stroke="#21262d" stroke-width="1"/>"##,
-        header_h,
-        total_w - 4,
-        header_h
-    ));
-    svg.push('\n');
-
-    // 1. Draw Cell Backgrounds
-    for y in 0..height {
-        for x in 0..width {
-            let cell = buffer.get(x, y);
-            let bg_color = match cell.bg {
-                ratatui::style::Color::DarkGray => "#21262d",
-                ratatui::style::Color::Gray => "#30363d",
-                ratatui::style::Color::Blue => "#1f6feb",
-                ratatui::style::Color::Red => "#da3633",
-                ratatui::style::Color::Green => "#238636",
-                ratatui::style::Color::Yellow => "#9e6a03",
-                ratatui::style::Color::Cyan => "#1b7c83",
-                ratatui::style::Color::Magenta => "#8957e5",
-                _ => continue,
-            };
-
-            let px = 8 + (x as u32) * char_w;
-            let py = header_h + (y as u32) * char_h;
-            svg.push_str(&format!(
-                r#"  <rect x="{}" y="{}" width="{}" height="{}" fill="{}"/>"#,
-                px, py, char_w, char_h, bg_color
-            ));
-            svg.push('\n');
-        }
-    }
-
-    // 2. Draw Cell Text Glyphs
-    for y in 0..height {
-        for x in 0..width {
-            let cell = buffer.get(x, y);
-            let symbol = cell.symbol();
-            if symbol.is_empty() || symbol == " " {
-                continue;
-            }
-
-            let fg_color = match cell.fg {
-                ratatui::style::Color::Red => "#ff7b72",
-                ratatui::style::Color::Green => "#7ee787",
-                ratatui::style::Color::Yellow => "#f2cc60",
-                ratatui::style::Color::Blue => "#79c0ff",
-                ratatui::style::Color::Magenta => "#d2a8ff",
-                ratatui::style::Color::Cyan => "#56d4dd",
-                ratatui::style::Color::Gray => "#8b949e",
-                ratatui::style::Color::DarkGray => "#484f58",
-                ratatui::style::Color::White => "#f0f6fc",
-                _ => "#c9d1d9",
-            };
-
-            let px = 8 + (x as u32) * char_w + 1;
-            let py = header_h + (y as u32) * char_h + 15;
-
-            let escaped = escape_xml(symbol);
-            svg.push_str(&format!(
-                r#"  <text x="{}" y="{}" fill="{}">{}</text>"#,
-                px, py, fg_color, escaped
-            ));
-            svg.push('\n');
-        }
-    }
-
-    svg.push_str("</svg>\n");
-
-    fs::write(output_path, svg)?;
-    Ok(())
-}
 
 fn create_populated_app_state() -> AppState {
     let mut app = AppState::new(OverlayMode::External);
